azure_function_proxy/requirements.txt
azure-functions>=1.20.0
youtube-transcript-api>=0.6.2
yt-dlp>=2024.1.0
